Moved b2048, bad words to server side
Refactoring so we can isolate the client more effectively with only the server having access to the twitter api.
Eventually the client will just take text as input and return a video/image over HTTP.
